+++ src/stories/views/components/article/components/podcast/podcast_episode_article.stories.js
@@ -1,10 +1,10 @@
-import podcast_player_episode from './podcast_player_episode.hbs'
+import podcast_episode_article from './podcast_episode_article.hbs'
 import podcastPlayerEpisodeJson from './fixtures/podcast_player_episode.json'
 
 const Template = ({ ...args }) => {
     // You can either use a function to create DOM elements or use a plain html string!
     // return `<span>${topline}</span>`;
-    return podcast_player_episode({ ...args })
+    return podcast_episode_article({ ...args })
 }
 
 export default {